feat(cli): split `respan setup` into tracing and gateway modes
Add `respan setup tracing` and `respan setup gateway` subcommands and turn
the interactive `respan setup` into a dispatcher that asks "Tracing or
Gateway?" before running the chosen flow. All three share one orchestrator.

- Introduce `SetupBaseCommand extends BaseCommand` to own the setup flow
  methods, keeping setup-only concerns off the root command class. Backed by
  three focused libs: `colors.ts` (ANSI constants), `agents.ts` (agent
  registry + detection), and `skill-content.ts` (skill markdown + assembly).
- `verifyApiKey(mode)` branches by mode: tracing keeps the demo OTLP trace to
  /api/v2/traces; gateway sends a real completion through /api/chat/completions
  with a 200 / 401-403 / 429 / 402-credits classifier (credits branch is
  provisional, pending a real out-of-credits response).
- Skill bundle now writes `tracing-setup.md` (renamed from setup.md) and a new
  `gateway-setup.md` containing the detect -> slug-map -> fetch-live-doc ->
  wire flow, with the explicit framework->slug map and a "no instrumentation
  packages" rule. SKILL.md routing repointed accordingly. Launch prompt is now
  mode-specific.
- `doctor.ts` switched to the shared `colors.ts` / `agents.ts` libs, dropping
  its duplicate ANSI constants and binary check.

`package.json` gains `@types/node` for the TS build.
