+++ javascript-sdks/respan-cli/src/lib/colors.ts
@@ -1,0 +1,11 @@
+// Shared ANSI color/style constants for setup-related commands.
+// Consolidated from the copies previously inlined in commands/setup/index.ts
+// and commands/setup/doctor.ts.
+
+export const PC = '\x1b[38;2;100;131;240m';
+export const RESET = '\x1b[0m';
+export const DIM = '\x1b[2m';
+export const GREEN = '\x1b[32m';
+export const BOLD = '\x1b[1m';
+export const RED = '\x1b[31m';
+export const YELLOW = '\x1b[33m';
